+++ components/BettingTable.tsx
@@ -41,19 +41,6 @@ export default function BettingTable({tableData, setUserData, apiKey, addBetsDon
         setUserData(updatedUserData);
     };
 
-    const handleMarketCorrectionTimeChange = (slug, time) => {
-        // Update the user data
-        const updatedUserData = tableData.map((row) => {
-            if (row.slug === slug) {
-                let newRow:userQuestion = row;
-                newRow.marketCorrectionTime = time;
-                return newRow;
-            }
-            return row;
-        });
-        setUserData(updatedUserData);
-    }
-
     const handleDeleteRow = (slug) => {
         const updatedData = [...tableData];
         const index = updatedData.findIndex(row => row.slug === slug);
